+++ CHANGELOG.md
@@ -8,8 +8,8 @@
 * methods in `Aspect` no longer return the aspect, use cascading operator to chain calls
 * improved type safety for `world.getManager` and `world.getSystem`, no longer takes a `Type` as parameter and uses
 generic methods instead (e.g. `world.getManager<TagManager>()` instead of `world.getManager(TagManager)`)
-* removed `Type` parameter in constructor of `Mapper`, change code from `Mapper<Position>(Position, world)` to
-`Mapper<Position>(world)`
+* removed `Type` parameter in constructor of `Mapper`, change code from `Mapper<Position>(Position, world)` to `Mapper<Position>(world)`
+* removed `uniqueId` from `Entity`
 ### Enhancements
 * `world.destroy()` for cleaning up `EntitySystem`s and `Manager`s